meson: Share common C source prefixes
gnu_source_prefix defines _GNU_SOURCE for compiler object functions.
The definition is universally available in the code base.

docs/devel/style.rst also says that the "qemu/osdep.h" header is
always included, so files included in the file is also universally
available in the code base.

Rename gnu_source_prefix to osdep_prefix, and add #include directives
that are referred by the users of gnu_source_prefix and contained in
qemu/osdep.h to safely de-duplicate #include directives.

getcpu is exceptional and uses sched_prefix, which appends
>#include <sched.h> to osdep_prefix. This is because cc.has_function()
returns false if the prefix has #include directives but the function is
not declared by the included files.

Signed-off-by: Akihiko Odaki <akihiko.odaki@daynix.com>